Fix eating and drinking verbs showing up after a short delay and making your verb UI bounce
