+++ Content.Shared/Burial/BurialSystem.cs
@@ -1,5 +1,4 @@
 using Content.Shared.ActionBlocker;
-using Content.Shared.Burial;
 using Content.Shared.Burial.Components;
 using Content.Shared.DoAfter;
 using Content.Shared.Interaction;
@@ -10,7 +9,7 @@ using Content.Shared.Storage.Components;
 using Content.Shared.Storage.EntitySystems;
 using Robust.Shared.Audio.Systems;
 
-namespace Content.Server.Burial.Systems;
+namespace Content.Shared.Burial;
 
 public sealed class BurialSystem : EntitySystem
 {
