+++ Content.Shared/EntityEffects/EffectConditions/InternalsCondition.cs
@@ -1,8 +1,7 @@
 using Content.Shared.Body.Components;
-using Content.Shared.EntityEffects;
 using Robust.Shared.Prototypes;
 
-namespace Content.Server.EntityEffects.EffectConditions;
+namespace Content.Shared.EntityEffects.EffectConditions;
 
 /// <summary>
 ///     Condition for if the entity is or isn't wearing internals.
