+++ Content.Shared/Ghost/Roles/Raffles/GhostRoleRaffleSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace Content.Server.Ghost.Roles.Raffles;
+﻿namespace Content.Shared.Ghost.Roles.Raffles;
 
 /// <summary>
 /// Defines settings for a ghost role raffle.
