+++ Content.Shared/Nutrition/Components/FlavorProfileComponent.cs
@@ -1,22 +1,24 @@
-namespace Content.Server.Nutrition.Components;
+using Robust.Shared.GameStates;
 
-[RegisterComponent]
+namespace Content.Shared.Nutrition.Components;
+
+[RegisterComponent, NetworkedComponent]
 public sealed partial class FlavorProfileComponent : Component
 {
     /// <summary>
     ///     Localized string containing the base flavor of this entity.
     /// </summary>
-    [DataField("flavors")]
+    [DataField]
     public HashSet<string> Flavors { get; private set; } = new();
 
     /// <summary>
     ///     Reagent IDs to ignore when processing this flavor profile. Defaults to nutriment.
     /// </summary>
-    [DataField("ignoreReagents")]
+    [DataField]
     public HashSet<string> IgnoreReagents { get; private set; } = new()
     {
         "Nutriment",
         "Vitamin",
-        "Protein"
+        "Protein",
     };
 }
