+++ Content.Shared/Nutrition/Components/UtensilComponent.cs
@@ -4,7 +4,7 @@ using Robust.Shared.GameStates;
 
 namespace Content.Shared.Nutrition.Components
 {
-    [RegisterComponent, NetworkedComponent, Access(typeof(SharedUtensilSystem))]
+    [RegisterComponent, NetworkedComponent, Access(typeof(UtensilSystem))]
     public sealed partial class UtensilComponent : Component
     {
         [DataField("types")]
