+++ Content.Shared/Nutrition/EntitySystems/FlavorProfileSystem.cs
@@ -1,12 +1,11 @@
-using Content.Server.Nutrition.Components;
+using System.Linq;
 using Content.Shared.CCVar;
 using Content.Shared.Chemistry.Components;
-using Content.Shared.Nutrition;
+using Content.Shared.Nutrition.Components;
 using Robust.Shared.Configuration;
 using Robust.Shared.Prototypes;
-using System.Linq;
 
-namespace Content.Server.Nutrition.EntitySystems;
+namespace Content.Shared.Nutrition.EntitySystems;
 
 /// <summary>
 ///     Deals with flavor profiles when you eat something.
