--- Content.Shared/Nutrition/EntitySystems/SharedUtensilSystem.cs
@@ -1,5 +1,0 @@
-namespace Content.Shared.Nutrition.EntitySystems;
-
-public abstract class SharedUtensilSystem : EntitySystem
-{
-}
