+++ Content.Shared/Storage/Components/MagnetPickupComponent.cs
@@ -1,6 +1,6 @@
 using Content.Shared.Inventory;
 
-namespace Content.Server.Storage.Components;
+namespace Content.Shared.Storage.Components;
 
 /// <summary>
 /// Applies an ongoing pickup area around the attached entity.
